requirements.txt
epc-encoding-utils